platform/x86: asus-wmi: Remove all ROG Ally CSEE hacks
Remove the hacks introduced by these commits:
- fc73fe3b8501 "platform/x86: asus-wmi: disable USB0 hub on ROG Ally before suspend"
- 6222853365c5 "platform/x86: asus-wmi: ROG Ally increase wait time, allow MCU powersave"
- 5e9a9192205f "platform/x86: asus-wmi: Add quirk for ROG Ally X"

The suspend issue the above was trying to fix has been corrected in MCU
firmware updates by ASUS, where the root cause of issues was a flag in
the MCU powersave path that was not being cleared on resume, leading to
every second resume leaving the MCU in a powered off state.

The fixed MCU firmware has been available since 2024/10/09.
Minimum versions:
- ROG Ally 1: v319
- ROG Ally X: v313

Signed-off-by: Luke D. Jones <luke@ljones.dev>